WIP: Pseudocode reverse engineering of NonIntegral.hs
Started writing down the imperative pseudocode equivalent of IOHKs exponentiation library NonIntegral.hs.
The idea is to end up with a _literate code_ pseudocode file that will fit the quickcheck property tests when implemented. This file, when done, should be enough for @AndrewWestberg to get started with an actual implementation. To ensure we have 1-to-1 compatibility with the Haskell library we will need to port the Quickcheck properties as well.